First commit
In first commit each module with the exception of PWM module are
implementing their functionality very well and both header and source
files are quite heavily commented.

In practice this means the voltage and current measurements from four
panels and battery can be observed on the LCD screen. Also each
measurement channel can be calibrated and adjusted through an adjustment
menu which is controlled with a button next to the screen.